requirements.txt
six==1.13.0
selenium==3.141.0
pyaml==19.12.0
pytz==2019.3